Extract the loading of configs to a dedicated module
This will allow multiple classes to be loaded and updated in the same
manner
